gcdmc/requirements.txt
google-cloud-datastore==2.4.0
phonenumbers==8.12.40
pylint==2.9.3
pytest==6.2.5
yapf==0.31.0
